Moved duplicateNodeCheck method to nodeset types
Added omitempty tag to servicesOverride
