+++ api/v1beta1/openstackdataplanenodeset_types.go
@@ -17,11 +17,16 @@ limitations under the License.
 package v1beta1
 
 import (
+	"fmt"
+
+	"golang.org/x/exp/slices"
+
 	infranetworkv1 "github.com/openstack-k8s-operators/infra-operator/apis/network/v1beta1"
 	condition "github.com/openstack-k8s-operators/lib-common/modules/common/condition"
 	baremetalv1 "github.com/openstack-k8s-operators/openstack-baremetal-operator/api/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 )
 
 // OpenStackDataPlaneNodeSetSpec defines the desired state of OpenStackDataPlaneNodeSet
@@ -203,4 +208,29 @@ type DataplaneAnsibleImageDefaults struct {
 	TelemetryCeilometerCompute string
 	TelemetryCeilometerIpmi    string
 	TelemetryNodeExporter      string
+}
+
+// duplicateNodeCheck checks the NodeSetList for pre-existing nodes. If the user is trying to redefine an
+// existing node, we will return an error and block resource creation.
+func (r *OpenStackDataPlaneNodeSetSpec) duplicateNodeCheck(nodeSetList *OpenStackDataPlaneNodeSetList) (errors field.ErrorList) {
+	existingNodeNames := make([]string, 0)
+	for _, existingNode := range nodeSetList.Items {
+		for _, node := range existingNode.Spec.Nodes {
+			existingNodeNames = append(existingNodeNames, node.HostName)
+			if node.Ansible.AnsibleHost != "" {
+				existingNodeNames = append(existingNodeNames, node.Ansible.AnsibleHost)
+			}
+		}
+	}
+
+	for _, newNodeName := range r.Nodes {
+		if slices.Contains(existingNodeNames, newNodeName.HostName) || slices.Contains(existingNodeNames, newNodeName.Ansible.AnsibleHost) {
+			errors = append(errors, field.Invalid(
+				field.NewPath("Spec").Child("nodes"),
+				newNodeName,
+				fmt.Sprintf("node already exists in the cluster: %s", newNodeName.HostName)))
+		}
+	}
+
+	return
 }
